requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
coloredlogs==15.0.1
discord.py==2.3.2
frozenlist==1.4.1
humanfriendly==10.0
idna==3.6
multidict==6.0.4
pyreadline3==3.4.1
python-dotenv==1.0.0
yarl==1.9.4
